Америка никогда не существовала
There is no such thing as american scoring. The only american thing in tournament chess is the american pairing system wihich uses some variables to calculate your opponent. One of these was called americanpoints in the original sources of pairtwo. This one was unfortunatly mistaken to really contain an american scoring